Update schedule.html with schedule table
This is the first HTML edition of the schedule table, including edits to this page's HTML and CSS. This was constructed and tested outside of this website, so the br codes in each of the cells may need to be adjusted to fit the spacing of the website. Additionally, if the CSS styling does not go through, the main.css document for this website may need to be checked for conflicting code, particularly that referring to table, tr, or td.